backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
numpy<2.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
chromadb==0.4.18
zhipuai==2.0.1
llama-parse==0.1.3
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional but recommended for better PDF parsing
pdfplumber==0.10.0
python-docx==1.1.0
